fix new leap payment

+++ app/src/main/java/com/app/smartpos/common/DeviceFactory/NewLandDevice.java
@@ -20,7 +20,7 @@ public class NewLandDevice implements Device{
         intent.putExtra(ThirdTag.CHANNEL_ID, "acquire");
         intent.putExtra(ThirdTag.TRANS_TYPE, 2);
         intent.putExtra(ThirdTag.OUT_ORDERNO, "12345");
-        intent.putExtra(ThirdTag.AMOUNT, total*100);
+        intent.putExtra(ThirdTag.AMOUNT, (long)(total*100.0));
         intent.putExtra(ThirdTag.INSERT_SALE, true);
         intent.putExtra(ThirdTag.RF_FORCE_PSW, true);
         return intent;
